moved debug to the ECS system and add simple info
