+++ Explorer/Assets/DCL/PluginSystem/Global/MultiplayerMovementPlugin.cs
@@ -6,6 +6,7 @@ using DCL.DebugUtilities;
 using DCL.Multiplayer.Movement.Settings;
 using DCL.Multiplayer.Movement.Systems;
 using DCL.Multiplayer.Profiles.Entities;
+using DCL.Multiplayer.Profiles.Tables;
 using Global.AppArgs;
 using System.Threading;
 using Utility;
@@ -24,15 +25,16 @@ namespace DCL.PluginSystem.Global
         private readonly ExposedTransform playerTransform;
         private readonly ProvidedAsset<MultiplayerDebugSettings> debugSettings;
         private readonly bool useCompression;
+        private readonly IReadOnlyEntityParticipantTable entityParticipantTable;
 
         private ProvidedAsset<MultiplayerMovementSettings> settings;
 
         private Entity? selfReplicaEntity;
-        private MultiplayerMovementDebug multiplayerMovementDebug;
+        private MultiplayerMovementDebugSystem multiplayerMovementDebugSystem;
 
         public MultiplayerMovementPlugin(IAssetsProvisioner assetsProvisioner, MultiplayerMovementMessageBus messageBus, IDebugContainerBuilder debugBuilder
-          , RemoteEntities remoteEntities, ExposedTransform playerTransform,
-            ProvidedAsset<MultiplayerDebugSettings> debugSettings, IAppArgs appArgs)
+          , RemoteEntities remoteEntities, ExposedTransform playerTransform, ProvidedAsset<MultiplayerDebugSettings> debugSettings, IAppArgs appArgs,
+            IReadOnlyEntityParticipantTable entityParticipantTable)
         {
             this.assetsProvisioner = assetsProvisioner;
             this.messageBus = messageBus;
@@ -40,12 +42,12 @@ namespace DCL.PluginSystem.Global
             this.remoteEntities = remoteEntities;
             this.playerTransform = playerTransform;
             this.debugSettings = debugSettings;
+            this.entityParticipantTable = entityParticipantTable;
             this.useCompression = appArgs.TryGetValue(COMPRESSION_ARG_FLAG, out string? compression) && compression == "true";
         }
 
         public void Dispose()
         {
-            multiplayerMovementDebug.Dispose();
             messageBus.Dispose();
             settings.Dispose();
         }
@@ -63,8 +65,7 @@ namespace DCL.PluginSystem.Global
             RemotePlayersMovementSystem.InjectToWorld(ref builder, settings.Value, settings.Value.CharacterControllerSettings);
             RemotePlayerAnimationSystem.InjectToWorld(ref builder, settings.Value.ExtrapolationSettings);
             CleanUpRemoteMotionSystem.InjectToWorld(ref builder);
-
-            multiplayerMovementDebug = new MultiplayerMovementDebug(builder.World, arguments.PlayerEntity, debugBuilder, remoteEntities, playerTransform, debugSettings.Value, settings.Value);
+            MultiplayerMovementDebugSystem.InjectToWorld(ref builder, arguments.PlayerEntity, debugBuilder, remoteEntities, playerTransform, debugSettings.Value, settings.Value, entityParticipantTable);
         }
     }
 }
